pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ace-experiment-framework"
version = "0.1.0"
description = "Automated Compute Efficiency Experimental Lab - Config-driven experimental framework for discovering compute optimizations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ACE Lab"}
]

dependencies = [
    "pyyaml>=6.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
